+++ zproject/backends.py
@@ -17,6 +17,7 @@ from typing import Any, Dict, List, Optional, Set, Tuple, Union
 
 from django_auth_ldap.backend import LDAPBackend, _LDAPUser
 import django.contrib.auth
+from django.contrib.auth import get_backends
 from django.contrib.auth.backends import RemoteUserBackend
 from django.conf import settings
 from django.core.exceptions import ValidationError
@@ -589,6 +590,24 @@ def sync_user_from_ldap(user_profile: UserProfile) -> bool:
             do_deactivate_user(user_profile)
         return False
     return True
+
+# Quick tool to test whether you're correctly authenticating to LDAP
+def query_ldap(**options: str) -> None:
+    email = options['email']
+    for backend in get_backends():
+        if isinstance(backend, LDAPBackend):
+            ldap_attrs = _LDAPUser(backend, backend.django_to_ldap_username(email)).attrs
+            if ldap_attrs is None:
+                print("No such user found")
+            else:
+                for django_field, ldap_field in settings.AUTH_LDAP_USER_ATTR_MAP.items():
+                    value = ldap_attrs.get(ldap_field, ["LDAP field not present", ])[0]
+                    if django_field == "avatar":
+                        if isinstance(value, bytes):
+                            value = "(An avatar image file)"
+                    print("%s: %s" % (django_field, value))
+                if settings.LDAP_EMAIL_ATTR is not None:
+                    print("%s: %s" % ('email', ldap_attrs[settings.LDAP_EMAIL_ATTR]))
 
 class DevAuthBackend(ZulipAuthMixin):
     """Allow logging in as any user without a password.  This is used for
